refac: change to eventhub
