+++ source/bronze/src/opengeh_bronze/infrastructure/streams/kafka_stream.py
@@ -1,4 +1,4 @@
-from pyspark.sql import DataFrame, SparkSession
+from pyspark.sql import SparkSession
 
 import opengeh_bronze.infrastructure.shared_helpers as shared_helpers
 from opengeh_bronze.domain.constants.database_names import DatabaseNames
@@ -17,7 +17,6 @@ class KafkaStream:
 
     def __init__(self) -> None:
         self.kafka_options = SubmittedTransactionsStreamSettings().create_kafka_options()
-        self.topic = SubmittedTransactionsStreamSettings().event_hub_instance
 
     def submit_transactions(self, spark: SparkSession) -> None:
         checkpoint_location = shared_helpers.get_checkpoint_path(
@@ -29,14 +28,4 @@ class KafkaStream:
             f"{DatabaseNames.bronze_database}.{TableNames.bronze_submitted_transactions_table}"
         )
 
-    def write_stream(
-        self,
-        dataframe: DataFrame,
-    ):
-        checkpoint_location = shared_helpers.get_checkpoint_path(
-            StorageAccountSettings().DATALAKE_STORAGE_ACCOUNT, StorageContainerNames.bronze, "processed_transactions"
-        )
 
-        dataframe.writeStream.format("kafka").options(**self.kafka_options).option("topic", self.topic).option(
-            "checkpointLocation", checkpoint_location
-        ).trigger(availableNow=True).start()
